requirements.txt
ultralytics>=8.3.0
streamlit>=1.30.0
opencv-python
pillow
pandas
numpy
